Web Inspector: Table selection should not require that rows be in the cache https://bugs.webkit.org/show_bug.cgi?id=190993 <rdar://problem/45613812>
Reviewed by Devin Rousso.

Source/WebInspectorUI:

Deselecting a table row shouldn't depend on it being in the cache. This
allows us to implement select all, and other kinds of bulk selection
actions that affect rows that haven't been revealed.

* UserInterface/Views/Table.js:
(WI.Table.prototype.deselectRow):

LayoutTests:

Add a test that a table row that is not in the cache can be selected/deselected.

* inspector/table/table-selection-expected.txt:
* inspector/table/table-selection.html:


Identifier: 206071@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@237817 268f45cc-cd09-0410-ab3c-d52691b4dbfc