Crash under DOMWindow::postMessageTimerFired() https://bugs.webkit.org/show_bug.cgi?id=191217 <rdar://problem/40888466>
Reviewed by Geoffrey Garen.

Source/WebCore:

Protect the frame in DOMWindow::postMessageTimerFired() before calling dispatchEvent() as dispatching the
event may cause JS to run and this JS may cause the frame to be destroyed, in which case we will crash
when trying to use the frame on the next line.

Test: fast/dom/Window/remove-frame-in-message-event-handler.html

* page/DOMWindow.cpp:
(WebCore::DOMWindow::postMessageTimerFired):

LayoutTests:

Add layout test coverage.

* fast/dom/Window/remove-frame-in-message-event-handler-expected.txt: Added.
* fast/dom/Window/remove-frame-in-message-event-handler.html: Added.


Identifier: 206079@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@237825 268f45cc-cd09-0410-ab3c-d52691b4dbfc