[iOS] WebGL leaks exact GPU type https://bugs.webkit.org/show_bug.cgi?id=191393 Source/WebCore:
Reviewed by Tim Horton.

The fully exposed renderer info shouldn't be necessary
on iOS, where the GPUs and drivers are consistent
enough that people shouldn't need to write code specifically
for them. Reduce the ability to fingerprint by simply
returning "Apple GPU".

The other option would have been to disable the extension,
but I think it might still be useful to know you're on
an iOS device.

Test: fast/canvas/webgl/hide-some-renderer-info.html

* html/canvas/WebGL2RenderingContext.cpp: Return "Apple GPU"
on iOS.
(WebCore::WebGL2RenderingContext::getParameter):
* html/canvas/WebGLRenderingContext.cpp: Ditto.
(WebCore::WebGLRenderingContext::getParameter):

LayoutTests:

<rdar://problem/45788297>

Reviewed by Tim Horton.

A test that checks that when the GPU vendor is Apple we're
returning a generic renderer name.

* fast/canvas/webgl/hide-some-renderer-info-expected.txt: Added.
* fast/canvas/webgl/hide-some-renderer-info.html: Added.

Identifier: 206177@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@237940 268f45cc-cd09-0410-ab3c-d52691b4dbfc