[Win] Compile Service Worker support https://bugs.webkit.org/show_bug.cgi?id=191409
Reviewed by Youenn Fablet.

Fix compilation errors when ENABLE(SERVICE_WORKER)
on Windows with clang-cl. Clang on dllexport
platforms does not support specifying the
dllexport on both a class and members of the class
and unistd.h isn't provided but also appeared to
not be used.

No new tests, should be covered by existing tests.

* workers/service/ServiceWorkerProvider.h:
* workers/service/context/SWContextManager.cpp:


Identifier: 206400@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@238196 268f45cc-cd09-0410-ab3c-d52691b4dbfc