[iOS] Do not get stuck in indeterminate content observation state. https://bugs.webkit.org/show_bug.cgi?id=191719
Reviewed by Simon Fraser.

Source/WebCore:

Reset the _WKContentChange flag to WKContentNoChange when the last observing timer is removed and we are in
the "can't decide yet if it's a hover or click" state.
This bug prevents us from firing click event when JS installs and removes the same set of timer(s) during mouse-move dispatch.

Test: fast/events/touch/ios/stuck-with-hover-state.html

* platform/ios/wak/WKContentObservation.cpp:
(WebThreadRemoveObservedContentModifier):

LayoutTests:

* fast/events/touch/ios/stuck-with-hover-state-expected.txt: Added.
* fast/events/touch/ios/stuck-with-hover-state.html: Added.


Identifier: 206468@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@238273 268f45cc-cd09-0410-ab3c-d52691b4dbfc